Break out some logic
